UPOD UI/UX Proposal
Hey everyone, this is a design that I think incorporates the ideas we had in our first front-end meeting. I think the interface should look minimalist [so long as it stays practical to our needs] in order to project a sort of clean polished look to it, something that's pleasing to the eye. Because of this I assume it will be easily navigable by nature, reducing any clutter the user has to face to only those elements that are of the most use to them. 

On another note, it seems to me that when the user reaches their intended page of the website, and they'd like to visit another page perhaps in another category afterwards, they can do so by navigating through the left sidebars already opened instead of the search bar. That would depend on them knowing what categories surround their topic, but hopefully with our designs the categories will be easy to navigate through.

Pros:
- Easy to navigate
- Efficient, no clutter
- Polished, clean look and feel
- Simplistic design for more straightforward organization

Cons:
- Sidebars opened after search query might consume valuable screen space. Maybe after a number of seconds they could be hidden again and the details of their search topic are then centered on the page.