Clickable footer hyperlinks added to Design Document
To remain consistent with the other documents' footer hyperlinks, I felt it was necessary to continue the standard here as well, although I am aware that I'm on the SQA team for this Design Document. The change I made was only to enhance the document's functionality on behalf of consistency, and does not reflect any SQA request made on the Design SQA Sheet, nor is it intended to change any of the document's content.