Removed unused code and added a "Back To Repository" button
Returning back to the repository from the index.html file is now as easy as 1 2 3 with this brand new "Back To Repository" button. Gone are the days of fumbling around the keyboard for the correct "back" navigation shortcut, or for locating the browser's "back" button which may very well consume seconds if not seconds. Feature is in beta status, pending approval.