Update UPOD Database SQL.sql
Need to add a table that holds every ID so we can add IDs to every table in increasing order without creating duplicates. No primary key because the table of max id does not require one.  I do not think we need to reference any of the ids directly as this table will only hold the current largest id number.