Added a download button to index.html
Turns out that github produces its own accessible link for downloading the repository, so I added it as a button to the index.html page which will open up a prompt of where you'd like to save it (b/c it goes straight to the github link for downloading the .zip file). If you have the link to the htmlpreview of index.html on hand [   http://htmlpreview.github.io/?https://github.com/LucianoAbbott/UPOD/blob/master/docs/index.html   ], then you could use it as a quick jumping pad to the documents, the front and back-end components, and to downloading the repository itself without needing to login to github!