Updated indexes to make them operational with UPOD local directory
Hopefully, this should facilitate a much easier mode of access to front and back-end components after downloading the repository. Additionally, as front-end components require first to be opened from the downloaded repository in order to properly display all their elements, the index.html page will help navigate to them with ease and with the assurance that all elements will display properly. One caveat is that the index links are only operational when accessed from the local file index.html within the downloaded UPOD repository. If accessed from the web, the page should show an alert prompting you to download the repository first.