Updated UPOD version number to 0.4.1
At the time of the previous version number update, UPOD was at 738 commits, and at the time of writing this, it is now at 980 commits. This version number update is pretty arbitrary, but to add a little structure, I've tried to put a little logic behind the version choice. With a difference of 242 commits between then and now, we could say that with a version setup of 0.1.0, every 20 commits is worth a thousandth digit, and by default every 80 commits will increment the hundredth digit (when we increment the 3rd digit 4 times, increment the 2nd digit and reset the third digit to 0) - counting from 0.1.1, we get 0.1.2, 0.1.3, 0.2.0, 0.2.1, 0.2.2, 0.2.3, 0.3.0, 0.3.1, 0.3.2, 0.3.3, 0.4.0, and then 0.4.1. Although this incrementation system doesn't differentiate small from big changes, just the number of commits, it at the very least gives us some idea of where we're at based on what we've done so far. I considered including all the commits before 0.1.0, but that would include all the changes we've made to the documents which may not accurately portray the status of UPOD itself. If anyone else has something they'd like to change, or even upend this entire version numbering method, please feel free to do so.