About icon added to top-right corner of index.html (for local file only)
It links to the local about.html file of the downloaded repository, and will only show if accessed from the local index inside the downloaded repository. Admittedly one could also click the about.html link in the drop-down menu of the front-end index, and so to counter that method I'd like to raise the justification that this new icon looks really cool. Although in retrospect, it seems the royal decree of styling completion for index.html was a tad premature.