SQA Sheets can now be opened from UPOD library without downloading
After doing a bit of searching, I came across this online excel viewer which can display local and web excel files. Interestingly, it can retrieve and display the SQA sheets from their github download links. The only drawbacks I found with this were that the loading times were a bit slow on my end, and on some rare occasions the website would stall in the middle of loading the spreadsheet up, but compared to redundantly downloading the excel sheets, I'd say it's an improvement.