Hyperlinked the in-line reference made to the edit page for easier accessibility
And it'll only show when opened from the local Progress Report.html file inside the repository, for other instances like from the web version it'll just show as "(as seen on the edit page)". I also changed the square brackets back to the round brackets because parenthetical remarks seemed best suited to the graceful arches of round brackets.